Move getEventById E2E tests to separate file
